requirements.txt
Django==3.2.25
djangorestframework==3.12.4
django-extensions==3.1.0
psycopg2-binary==2.8.4
django-filter==2.4.0
factory-boy==3.1.0
celery[rabbitmq]==5.3.0
django-celery-beat==2.2.1
pytest-django==4.4.0
djangorestframework-simplejwt==5.3.1
requests==2.31.0